requirements.txt
fastapi[standard]==0.115.2
requests==2.32.3
python-dateutil